Create main2.cpp
Pull requests are disabled so I am uploading a new file instead of amending one. This copy of main.cpp has a revamped menu code to offer protection against erroneous input. Using a switch statement limits the cases in the "choice" variable to only 1-6 integers. The default case is executed if any other characters are inputted an notifies the user of this error. Also added an inFile.close(); to the exit option to make sure the program properly closes the .txt input file upon returning from execution.